+++ src/routes/g/[game]/[type]/play/+page.ts
@@ -30,6 +30,7 @@ export const load: PageLoad = async ({ params, url, fetch }) => {
 		entityType: params.type,
 		typeLabel: type.label,
 		editorFirst,
+		hasSheet: !!type.sheetComponent,
 		saved
 	};
 };